add solution strategies for problem solutions
