+++ org.eclipse.winery.model.adaptation/pom.xml
