+++ org.eclipse.winery.repository.rest/src/main/java/org/eclipse/winery/repository/rest/resources/MainResource.java
@@ -38,9 +38,8 @@ import org.eclipse.winery.repository.importing.CsarImportOptions;
 import org.eclipse.winery.repository.importing.CsarImporter;
 import org.eclipse.winery.repository.importing.ImportMetaInformation;
 import org.eclipse.winery.repository.rest.RestUtils;
-import org.eclipse.winery.repository.rest.resources.testrefinementmodels.TestRefinementModelsResource;
-import org.eclipse.winery.repository.rest.resources.admin.AdminTopResource;
 import org.eclipse.winery.repository.rest.resources.API.APIResource;
+import org.eclipse.winery.repository.rest.resources.admin.AdminTopResource;
 import org.eclipse.winery.repository.rest.resources.compliancerules.ComplianceRulesResource;
 import org.eclipse.winery.repository.rest.resources.entitytemplates.artifacttemplates.ArtifactTemplatesResource;
 import org.eclipse.winery.repository.rest.resources.entitytemplates.policytemplates.PolicyTemplatesResource;
@@ -55,6 +54,7 @@ import org.eclipse.winery.repository.rest.resources.entitytypes.requirementtypes
 import org.eclipse.winery.repository.rest.resources.imports.ImportsResource;
 import org.eclipse.winery.repository.rest.resources.patternrefinementmodels.PatternRefinementModelsResource;
 import org.eclipse.winery.repository.rest.resources.servicetemplates.ServiceTemplatesResource;
+import org.eclipse.winery.repository.rest.resources.testrefinementmodels.TestRefinementModelsResource;
 import org.eclipse.winery.repository.rest.resources.yaml.YAMLParserResource;
 
 import com.sun.jersey.core.header.FormDataContentDisposition;
@@ -76,6 +76,7 @@ import org.slf4j.LoggerFactory;
 @Path("/")
 public class MainResource {
     private static final Logger LOGGER = LoggerFactory.getLogger(MainResource.class);
+
     @Path("API/")
     public APIResource api() {
         return new APIResource();
@@ -220,7 +221,7 @@ public class MainResource {
         }
         if (importMetaInformation.errors.isEmpty()) {
             if (options.isValidate()) {
-                
+
                 return Response.ok(importMetaInformation, MediaType.APPLICATION_JSON).build();
             } else if (Objects.nonNull(importMetaInformation.entryServiceTemplate)) {
                 URI url = uriInfo.getBaseUri().resolve(RestUtils.getAbsoluteURL(importMetaInformation.entryServiceTemplate));
