+++ org.eclipse.winery.repository.rest/src/main/java/org/eclipse/winery/repository/rest/resources/apiData/RefinementElementApiData.java
@@ -17,7 +17,7 @@ package org.eclipse.winery.repository.rest.resources.apiData;
 import java.util.List;
 
 import org.eclipse.winery.common.ids.definitions.ServiceTemplateId;
-import org.eclipse.winery.model.substitution.refinement.RefinementCandidate;
+import org.eclipse.winery.model.adaptation.substitution.refinement.RefinementCandidate;
 import org.eclipse.winery.model.tosca.TTopologyTemplate;
 
 public class RefinementElementApiData {
